Add Header component
+ some component updates
