+++ webui/src/components/LoadingSpinner.vue
@@ -1,3 +1,9 @@
+<template>
+  <div class="text-gray-300 dark:text-red-300 pt-24 pb-12">
+    <IconSpinnerVue class="w-5 animate-spin mx-auto" />
+  </div>
+</template>
+
 <script>
 import IconSpinnerVue from './icons/IconSpinner.vue';
 
@@ -5,7 +11,3 @@ export default {
   components: { IconSpinnerVue },
 };
 </script>
-
-<template>
-  <IconSpinnerVue class="w-5 animate-spin mx-auto" />
-</template>
