+++ webui/src/store/store.js
@@ -1,5 +1,6 @@
 import API from '@/services/api';
 import { ref, reactive } from 'vue';
+import sha256 from 'crypto-js/sha256';
 
 import { defineStore } from 'pinia';
 
@@ -33,7 +34,7 @@ export const useStore = defineStore('store', () => {
         const session = await api.getSession();
         authenticated.value = session.authenticated;
         requiresPassword.value = session.requiresPassword;
-        // return this.refresh();
+        return this.refresh();
       })
       .catch((err) => {
         console.log(err.message || err.toString());
@@ -64,9 +65,7 @@ export const useStore = defineStore('store', () => {
     api
       .createClient({ name: clientCreateName.value })
       .catch((err) => alert(err.message || err.toString()))
-      .finally
-      // () => refresh().catch(console.error)
-      ();
+      .finally(() => refresh().catch(console.error));
 
     clientCreateShowModal.value = null;
   }
@@ -75,11 +74,65 @@ export const useStore = defineStore('store', () => {
     api
       .deleteClient({ clientId: client.id })
       .catch((err) => alert(err.message || err.toString()))
-      .finally
-      // () => refresh().catch(console.error)
-      ();
+      .finally(() => refresh().catch(console.error));
 
     clientToDelete.value = null;
+  }
+
+  async function refresh({ updateCharts = false } = {}) {
+    if (!authenticated.value) return;
+
+    try {
+      const clientsData = await api.getClients();
+      clients.value = clientsData.map((client) => {
+        if (client.name.includes('@') && client.name.includes('.')) {
+          client.avatar = `https://www.gravatar.com/avatar/${sha256(client.name)}?d=blank`;
+        }
+
+        if (!clientsPersist[client.id]) {
+          clientsPersist[client.id] = {};
+          clientsPersist[client.id].transferRxHistory = Array(50).fill(0);
+          clientsPersist[client.id].transferRxPrevious = client.transferRx;
+          clientsPersist[client.id].transferTxHistory = Array(50).fill(0);
+          clientsPersist[client.id].transferTxPrevious = client.transferTx;
+        }
+
+        // Debug
+        client.transferRx = clientsPersist[client.id].transferRxPrevious + Math.random() * 1000;
+        client.transferTx = clientsPersist[client.id].transferTxPrevious + Math.random() * 1000;
+        client.latestHandshakeAt = new Date('2024-03-20');
+        updateCharts = true; // DEV TODO: Update. Get from settings
+
+        if (updateCharts) {
+          clientsPersist[client.id].transferRxCurrent =
+            client.transferRx - clientsPersist[client.id].transferRxPrevious;
+          clientsPersist[client.id].transferRxPrevious = client.transferRx;
+          clientsPersist[client.id].transferTxCurrent =
+            client.transferTx - clientsPersist[client.id].transferTxPrevious;
+          clientsPersist[client.id].transferTxPrevious = client.transferTx;
+
+          clientsPersist[client.id].transferRxHistory.push(clientsPersist[client.id].transferRxCurrent);
+          clientsPersist[client.id].transferRxHistory.shift();
+
+          clientsPersist[client.id].transferTxHistory.push(clientsPersist[client.id].transferTxCurrent);
+          clientsPersist[client.id].transferTxHistory.shift();
+        }
+
+        client.transferTxCurrent = clientsPersist[client.id].transferTxCurrent;
+        client.transferRxCurrent = clientsPersist[client.id].transferRxCurrent;
+
+        client.transferTxHistory = clientsPersist[client.id].transferTxHistory;
+        client.transferRxHistory = clientsPersist[client.id].transferRxHistory;
+        client.transferMax = Math.max(...client.transferTxHistory, ...client.transferRxHistory);
+
+        client.hoverTx = clientsPersist[client.id].hoverTx;
+        client.hoverRx = clientsPersist[client.id].hoverRx;
+
+        return client;
+      });
+    } catch (error) {
+      console.log('Something went wrong', error);
+    }
   }
 
   return {
@@ -97,5 +150,6 @@ export const useStore = defineStore('store', () => {
     logout,
     createClient,
     deleteClient,
+    refresh,
   };
 });
